Bot: va bay bo cuc TProcessData truoc khi mo duong Core -> Goddess
Doi DUNG MOT TU o Sources\Core\Src\KProtocol.h:1643  size_t -> unsigned int.

BOI CANH: cay nay co HAI tep KProtocol.h dung CHUNG guard KPROTOCOL_H
(Headers\KProtocol.h va Sources\Core\Src\KProtocol.h) nen ban nao include truoc thi ban kia
bi bo qua AM THAM. Chung khac nhau dung mot truong: nDataLen la unsigned int vs size_t.
TProcessData la GOI TIN QUA MANG giua cac tien trinh KHAC BITNESS - do PE header ban dang
chay: Goddess.exe/Bishop.exe = Win32, GameServer.exe/CoreServer.dll = x64.

BO CUC (ca hai tep deu o #pragma pack(1) dong 18-19 nen khong padding; da kiem chung bang
chinh cl.exe 14.44 cho ca x86 lan x64):
  Headers  (unsigned int): Win32 = 11 byte, x64 = 11 byte
  Core\Src (size_t)      : Win32 = 11 byte, x64 = 15 byte  <-- lech
  offset ulIdentity nhay 5 -> 9, bLeave 9 -> 13.

DUONG DANG CHAY HOM NAY KHONG HONG - da truy tan noi:
  Goddess    : ClientNode.cpp -> IDBRoleServer.h:12 -> Lib\S3DBInterface.h:9 -> Headers  = 11
  Bishop     : GamePlayer.cpp:8 truc tiep -> Headers = 11
  GameServer : Lib\S3DBInterface.h -> Headers = 11. (Dong #include "KCore.h" o KSOServer.cpp:19
               - duong duy nhat co the keo ban size_t vao - nam TRONG #ifdef _STANDALONE DANG TAT.)
  Core       : thay ban 15 byte NHUNG hom nay KHONG dung TProcessData o bat ky dong nao
               (grep toan Sources\Core chi trung dung tep KProtocol.h dinh nghia no).
=> nhat quan 11 byte tren toan he. Khong phai loi dang song.

VI SAO VAN PHAI VA: ma bot se nam TRONG Core - noi thay 15 byte - roi ban thang sang Goddess
doc 11 byte. GameServer chi chuyen tiep (const void*, int) nen khong the phat hien. Hau qua:
Goddess doc ulIdentity = dword cao cua nDataLen = 0, ten tai khoan lech 4 byte ->
GetRoleListOfAccount that bai -> tra 0 nhan vat -> bot khong bao gio duoc tao, KHONG co
mot dong bao loi nao.

VI SAO CHON CACH NAY: Core khong dung struct nay nen doi kieu la 0 thay doi hanh vi cho moi
tien trinh dang chay, chi can build lai Core. Hai cach khac deu te hon:
 - dat guard khac nhau: PHA BUILD NGAY (283 kieu trung ten giua hai tep)
 - #pragma pack: VO TAC DUNG (pack(1) da bat san; pack khong doi duoc kich thuoc mot thanh vien)

Goc re: cay tham khao D:\USVOLAM (bot chay that) chi co MOT ban KProtocol.h duy nhat.
Ban fork trong Sources\Core\Src cua cay nay chinh la nguon goc bay.

Build: Server Release|x64 PASS va Client Release|Win32 PASS (0 loi ca hai).
Byte cao KProtocol.h 11063 -> 11063, FFFD 0.

+++ Sources/Core/Src/KProtocol.h
@@ -1637,10 +1637,21 @@ typedef struct
 
 //ÒÆ×ÔRoleDBManager/kroledbheader.h
 //ÓÃÀ´Ìæ»»ÉÏÃæµÄROLE_LIST_SYNC,ROLE_LIST_SYNC½á¹¹²»ÔÙÐèÒªÁË
+// LUU Y BO CUC BYTE - dung doi kieu cac truong duoi day.
+//   Co HAI tep KProtocol.h trong cay nay va chung dung CHUNG guard KPROTOCOL_H
+//   (Headers\KProtocol.h va tep nay), nen ban nao include truoc thi ban kia bi bo qua.
+//   TProcessData la GOI TIN QUA MANG giua cac tien trinh KHAC BITNESS:
+//     Goddess.exe / Bishop.exe = Win32,  GameServer.exe / CoreServer.dll = x64.
+//   Ca hai tep deu o #pragma pack(1) (dong 18-19) nen khong co padding; nhung neu
+//   nDataLen la size_t thi tren x64 no thanh 8 byte, day ulIdentity tu offset 5 sang 9
+//   va sizeof tu 11 len 15. Goddess (Win32) van doc theo 11 byte => doc nham ulIdentity,
+//   ten tai khoan lech 4 byte, GetRoleListOfAccount that bai va KHONG bao loi.
+//   => Phai giu unsigned int de moi tien trinh cung thay 11 byte.
+//   (Cay tham khao D:\USVOLAM chi co MOT ban KProtocol.h nen khong dinh bay nay.)
 struct TProcessData
 {
 	unsigned char	nProtoId;
-	size_t			nDataLen;//TRoleNetMsgÊ±±íÊ¾¸ÃBlockµÄÊµ¼ÊÊý¾Ý³¤¶È,TProcessDataÊ±±íÊ¾StreamµÄÊµ¼ÊÊý¾Ý³¤¶È
+	unsigned int		nDataLen;//TRoleNetMsgÊ±±íÊ¾¸ÃBlockµÄÊµ¼ÊÊý¾Ý³¤¶È,TProcessDataÊ±±íÊ¾StreamµÄÊµ¼ÊÊý¾Ý³¤¶È
 	unsigned long	ulIdentity;
 	bool			bLeave;
 	char			pDataBuffer[1];//Êµ¼ÊµÄÊý¾Ý
